# sql: initialize ULID entropy lazily