--- a/pkg/sql/BUILD.bazel
+++ b/pkg/sql/BUILD.bazel
@@ -589,7 +589,6 @@ go_library(
         "//pkg/util/tracing/tracingpb",
         "//pkg/util/tsearch",
         "//pkg/util/uint128",
-        "//pkg/util/ulid",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_apd_v3//:apd",
         "@com_github_cockroachdb_errors//:errors",