This commit is similar to 9e1fdbaf585154640f73e9440ea344e589406016 in spirit and applies the "lazy allocation" optimization to ULID entropy. We now create `ulid.MonotonicReader` the first time `gen_random_ulid` is invoked. Additionally, we now use the same RNG as the one that powers `random` builtin as the entropy source for ULID generation.

Epic: CRDB-39063.

Release note: None