@yuzefovich 2024-06-06
Impact on a microbenchmark:
```
name                        old time/op    new time/op    delta
Infrastructure/n1/r1-8        15.3µs ± 1%    14.1µs ± 2%   -7.94%  (p=0.000 n=10+10)
Infrastructure/n1/r100-8      54.4µs ± 1%    52.8µs ± 1%   -2.85%  (p=0.000 n=9+9)
Infrastructure/n1/r10000-8    3.79ms ± 2%    3.75ms ± 1%     ~     (p=0.075 n=10+10)
Infrastructure/n3/r1-8         523µs ± 5%     490µs ± 5%   -6.25%  (p=0.000 n=10+10)
Infrastructure/n3/r100-8       691µs ± 5%     687µs ± 6%     ~     (p=0.481 n=10+10)
Infrastructure/n3/r10000-8    14.5ms ± 8%    14.3ms ±12%     ~     (p=0.529 n=10+10)

name                        old alloc/op   new alloc/op   delta
Infrastructure/n1/r1-8        22.1kB ± 1%    17.8kB ± 1%  -19.39%  (p=0.000 n=9+9)
Infrastructure/n1/r100-8      67.9kB ± 0%    63.8kB ± 1%   -5.99%  (p=0.000 n=8+9)
Infrastructure/n1/r10000-8    4.54MB ± 0%    4.53MB ± 0%   -0.16%  (p=0.006 n=8+9)
Infrastructure/n3/r1-8         147kB ±13%     131kB ± 1%  -10.89%  (p=0.000 n=8+8)
Infrastructure/n3/r100-8       372kB ± 0%     369kB ±12%     ~     (p=0.050 n=9+9)
Infrastructure/n3/r10000-8    21.5MB ± 0%    21.5MB ± 0%     ~     (p=0.200 n=8+9)

name                        old allocs/op  new allocs/op  delta
Infrastructure/n1/r1-8           112 ± 1%       108 ± 0%   -3.14%  (p=0.000 n=10+8)
Infrastructure/n1/r100-8         225 ± 0%       223 ± 3%     ~     (p=0.066 n=7+9)
Infrastructure/n1/r10000-8     10.3k ± 0%     10.3k ± 0%     ~     (p=0.409 n=8+9)
Infrastructure/n3/r1-8         1.49k ± 9%     1.46k ± 0%   -2.14%  (p=0.000 n=8+8)
Infrastructure/n3/r100-8       2.11k ± 0%     2.18k ±17%     ~     (p=0.398 n=9+9)
Infrastructure/n3/r10000-8     53.4k ± 0%     53.5k ± 0%     ~     (p=0.606 n=8+9)
```